requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0